requirements.txt
torch
torchvision
Pillow
numpy
tqdm
tensorboard